feat: update to next@16 Fixes #282
