+++ package.json
@@ -41,7 +41,7 @@
     "@dnd-kit/sortable": "^8.0.0",
     "@faker-js/faker": "^8.4.1",
     "@geoman-io/leaflet-geoman-free": "^2.19.3",
-    "@graphql-yoga/plugin-persisted-operations": "^3.17.1",
+    "@graphql-yoga/plugin-persisted-operations": "^3.21.2",
     "@heroui/react": "^2.8.5",
     "@hookform/resolvers": "^3.3.4",
     "@iconify/react": "^5.2.1",
@@ -69,13 +69,10 @@
     "dom-to-image": "^2.6.0",
     "echarts": "^6.0.0",
     "echarts-for-react": "^3.0.6",
-    "embla-carousel-autoplay": "^8.1.6",
-    "embla-carousel-class-names": "^8.6.0",
-    "embla-carousel-react": "^8.1.6",
     "graphql": "^16.11.0",
     "graphql-scalars": "^1.23.0",
     "graphql-sse": "^2.5.3",
-    "graphql-yoga": "^5.15.1",
+    "graphql-yoga": "^5.21.2",
     "gridstack": "^12.6.0",
     "http-status-codes": "^2.3.0",
     "immer": "^10.1.1",
@@ -85,8 +82,8 @@
     "lexical": "^0.27.1",
     "lexical-beautiful-mentions": "^0.1.47",
     "motion": "^12.23.12",
-    "next": "^15.5",
-    "next-recaptcha-v3": "^1.5.2",
+    "next": "16.2.9",
+    "next-recaptcha-v3": "^1.5.3",
     "next-themes": "^0.4.6",
     "nodemailer": "^8.0.5",
     "object-hash": "^3.0.0",
@@ -94,8 +91,8 @@
     "pino": "^8.19.0",
     "pino-pretty": "^11.0.0",
     "query-string": "^9.1.1",
-    "react": "19.1.1",
-    "react-dom": "19.1.1",
+    "react": "19.2.7",
+    "react-dom": "19.2.7",
     "react-hook-form": "^7.61.1",
     "react-leaflet": "^5.0.0",
     "react-multi-email": "^1.0.25",
@@ -129,15 +126,15 @@
     "@types/node": "^20",
     "@types/nodemailer": "^6.4.17",
     "@types/object-hash": "^3.0.6",
-    "@types/react": "19.1.9",
-    "@types/react-dom": "19.1.7",
+    "@types/react": "19.2.17",
+    "@types/react-dom": "19.2.3",
     "@typescript-eslint/eslint-plugin": "^8.49.0",
     "@typescript-eslint/parser": "^8.49.0",
     "babel-jest": "^30.2.0",
     "commander": "^13.1.0",
     "dotenv-cli": "^7.4.2",
     "eslint": "^9.39.4",
-    "eslint-config-next": "^16.2.4",
+    "eslint-config-next": "16.2.9",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-jest": "^29.5.0",
     "eslint-plugin-playwright": "^2.4.0",
@@ -153,15 +150,15 @@
     "prettier-plugin-tailwindcss": "^0.7.2",
     "prisma": "6",
     "react-email": "^5.2.8",
-    "semantic-release": "^25.0.3",
+    "semantic-release": "^25.0.5",
     "shx": "^0.4.0",
     "tailwindcss": "^4.1.17",
     "ts-node": "^10.9.2",
     "typescript": "^5"
   },
   "resolutions": {
-    "@types/react": "19.1.9",
-    "@types/react-dom": "19.1.7"
+    "@types/react": "19.2.17",
+    "@types/react-dom": "19.2.3"
   },
   "packageManager": "yarn@4.12.0"
 }
