--- src/app/view/landing/feature-overview-as-slideshow/carousel/styles.module.css
@@ -1,8 +1,0 @@
-/**
- * .is-snapped is coming from the Embla ClassNames plugin
-
- * See: https://www.embla-carousel.com/plugins/class-names/
- */
-.slide:not(:global(.is-snapped)) {
-  opacity: 0.16;
-}
