--- src/app/view/landing/feature-overview-as-slideshow/lightbox/index.tsx
@@ -1,82 +1,0 @@
-'use client';
-import { cn } from '@heroui/react';
-import { type UseDisclosureReturn } from '@heroui/use-disclosure';
-import useEmblaCarousel from 'embla-carousel-react';
-import React from 'react';
-import { Button } from '~/view/base/button';
-import { Icon } from '~/view/base/icon';
-import { Modal } from '~/view/base/modal';
-import { type ScreenshotEntry } from '../../feature-overview-image-list';
-import { useEmblaApi } from '../use-embla-api';
-import { Slide } from './slide';
-
-interface Props {
-  className?: string;
-  imageList: ScreenshotEntry[];
-  disclosure: UseDisclosureReturn;
-  startIndex: number;
-  onClose?: (imageIndex: number) => void;
-}
-
-export const Lightbox: React.FC<Props> = ({
-  className,
-  imageList,
-  disclosure,
-  startIndex,
-  onClose,
-}) => {
-  const [emblaRef, emblaApi] = useEmblaCarousel({ loop: true, startIndex }, []);
-  const { scrollNext, scrollPrev, selectedIndex } = useEmblaApi(emblaApi);
-
-  const modalClose = React.useCallback(() => {
-    disclosure.onClose();
-    onClose?.(selectedIndex);
-  }, [disclosure, onClose, selectedIndex]);
-
-  return (
-    <Modal
-      isOpen={disclosure.isOpen}
-      size="full"
-      scrollBehavior="inside"
-      onClose={modalClose}
-    >
-      <Modal.Content>
-        {(onModalClose) => {
-          return (
-            <>
-              {/**
-               * Put in an empty header so ModalBody that is holding the image won't hide the
-               * modal close button
-               */}
-              <Modal.Header />
-              <Modal.Body className="justify-center p-0">
-                <div
-                  className={cn(className, 'overflow-x-hidden')}
-                  ref={emblaRef}
-                >
-                  <div className="flex">
-                    {imageList.map((slide) => (
-                      <Slide key={slide.alt} {...slide} />
-                    ))}
-                  </div>
-                </div>
-              </Modal.Body>
-              <Modal.Footer className="flex items-center justify-center gap-3">
-                <Button
-                  isIconOnly
-                  startContent={<Icon icon="back" />}
-                  onPress={() => scrollPrev()}
-                />
-                <Button
-                  isIconOnly
-                  startContent={<Icon className="rotate-180" icon="back" />}
-                  onPress={() => scrollNext()}
-                />
-              </Modal.Footer>
-            </>
-          );
-        }}
-      </Modal.Content>
-    </Modal>
-  );
-};
