+++ src/app/view/landing/index.tsx
@@ -1,7 +1,6 @@
 'use client';
 import React from 'react';
 import { FeatureOverviewAsGallery } from './feature-overview-as-gallery';
-import { FeatureOverviewAsSlideshow } from './feature-overview-as-slideshow';
 import { Hero } from './hero';
 import { SignUpReason } from './sign-up-reason';
 
@@ -14,7 +13,6 @@ export const Landing: React.FC<Props> = ({ className }) => {
     <div className={className}>
       <Hero />
       <FeatureOverviewAsGallery />
-      {/* <FeatureOverviewAsSlideshow /> */}
       <SignUpReason />
     </div>
   );
