--- src/middleware.ts
@@ -1,8 +1,0 @@
-import { authMiddleware } from './middlewares/auth';
-import { chain } from './middlewares/chain';
-
-/** Middle wares are executed in order until it returns a response */
-export default chain([
-  // Verify authentication
-  authMiddleware,
-]);
